~300 null safety errors squashed, the update marches on

+++ web/scripts/engine/rules.dart
@@ -12,5 +12,5 @@ class RuleSet {
     double gravity = 300;
 
     /// Maximum life value to start a level with
-    int maxLife = 100;
+    double maxLife = 100;
 }
